+++ apps/example-expo/app/example-bottom-sheet-with-keyboard.tsx
@@ -1,6 +1,7 @@
-import { Fragment, useState } from 'react'
-import { Button, StyleSheet, Text, View } from 'react-native'
+import { Fragment, useMemo, useState } from 'react'
+import { Button, Platform, StyleSheet, Text, View } from 'react-native'
 import { TextInput } from 'react-native-gesture-handler'
+import { useSafeAreaInsets } from 'react-native-safe-area-context'
 import {
   Backdrop,
   BottomSheet,
@@ -13,7 +14,22 @@ import {
 import { Portal } from 'react-native-universe-portal'
 
 export default function ExampleBottomSheetWithKeyboard() {
+  const { top, bottom } = useSafeAreaInsets()
   const [isOpenA, setIsOpenA] = useState(false)
+
+  const keyboardOffset = useMemo(() => {
+    const baseOffset = 20
+
+    if (Platform.OS === 'ios') {
+      return baseOffset
+    }
+
+    if (Platform.OS === 'android') {
+      return baseOffset + top + bottom
+    }
+  }, [bottom, top])
+
+  // MARK: Renderers
 
   const renderContent = () => {
     return (
@@ -62,7 +78,7 @@ export default function ExampleBottomSheetWithKeyboard() {
               </BottomSheetScrollView>
             </BottomSheet>
 
-            <KeyboardExpander />
+            <KeyboardExpander keyboardOffset={keyboardOffset} />
           </BottomSheetPresenter>
         </SheetStackItem>
       </Portal>